Improve hashing
