+++ lib/core/figbird.js
@@ -1,4 +1,4 @@
-import { hashObject } from './helpers'
+import { hashObject } from './hash'
 
 /**
     Usage:
