+++ test/snapshots/figbird.test.js.md
@@ -23,14 +23,14 @@ Generated by [AVA](https://avajs.dev).
         },
         itemQueryIndex: {
           1: [
-            'q/QbxqDzsAAAA=',
+            'q/BTSlpA==',
           ],
           2: [
-            'q/QbxqDzsAAAA=',
+            'q/BTSlpA==',
           ],
         },
         queries: {
-          'q/QbxqDzsAAAA=': {
+          'q/BTSlpA==': {
             config: {
               fetchPolicy: 'swr',
               realtime: 'merge',
@@ -45,7 +45,7 @@ Generated by [AVA](https://avajs.dev).
               serviceName: 'notes',
             },
             pending: false,
-            queryId: 'q/QbxqDzsAAAA=',
+            queryId: 'q/BTSlpA==',
             state: {
               data: [
                 {
